refactor(Viewer): Move hocs to ./hoc folder
