+++ react/Viewer/Footer/ForwardButton.jsx
@@ -10,7 +10,7 @@ import ReplyIcon from '../../Icons/Reply'
 import ShareIosIcon from '../../Icons/ShareIos'
 import Button from '../../Buttons'
 import Alerter from '../../Alerter'
-import { withViewerLocales } from '../withViewerLocales'
+import { withViewerLocales } from '../hoc/withViewerLocales'
 import { exportFilesNative } from './helpers'
 
 const {
