+++ react/Viewer/ViewersByFile/NoNetworkViewer.jsx
@@ -4,7 +4,7 @@ import Icon from '../../Icon'
 import CloudBrokenIcon from '../../Icons/CloudBroken'
 import Button from '../../Button'
 
-import { withViewerLocales } from '../withViewerLocales'
+import { withViewerLocales } from '../hoc/withViewerLocales'
 
 import styles from './styles.styl'
