+++ react/Viewer/ViewersByFile/OnlyOfficeViewer.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import Button from '../../Button'
 import { FileDoctype } from '../../proptypes'
 
-import { withViewerLocales } from '../withViewerLocales'
+import { withViewerLocales } from '../hoc/withViewerLocales'
 import NoViewer from '../NoViewer'
 
 const OnlyOfficeViewer = ({ file, onlyOfficeOpener, t }) => {
