+++ react/Viewer/ViewersByFile/TextViewer.jsx
@@ -8,7 +8,7 @@ import { withClient, models } from 'cozy-client'
 import { FileDoctype } from '../../proptypes'
 
 import ViewerSpinner from '../ViewerSpinner'
-import withFileUrl from '../withFileUrl'
+import withFileUrl from '../hoc/withFileUrl'
 import NoViewer from '../NoViewer'
 import { isFileEncrypted } from '../helpers'
