+++ react/Viewer/ViewersByFile/VideoViewer.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import withFileUrl from '../withFileUrl'
+import withFileUrl from '../hoc/withFileUrl'
 
 import styles from './styles.styl'
