+++ react/Viewer/hoc/withViewerLocales.jsx
@@ -1,0 +1,11 @@
+import withLocales from '../../I18n/withLocales'
+
+import en from '../locales/en.json'
+import fr from '../locales/fr.json'
+
+export const locales = {
+  en,
+  fr
+}
+
+export const withViewerLocales = withLocales(locales)
